Adding more refactoring changes!
